feat: sync new displays with primary screen
