+++ desktop video/desktop video/ContentView.swift
@@ -18,6 +18,8 @@ class AppState: ObservableObject {
     @Published var lastVolume: Float = 1.0
     @Published var lastStretchToFill: Bool = true
     @Published var currentMediaURL: String?
+    /// 当前主控制器对应的屏幕 ID
+    @Published var primaryScreenID: CGDirectDisplayID?
 }
 
 class ScreenObserver: ObservableObject {
@@ -27,6 +29,7 @@ class ScreenObserver: ObservableObject {
     private var previousScreens: [NSScreen] = NSScreen.screens
 
     init() {
+        AppState.shared.primaryScreenID = screens.first?.dv_displayID
         observer = NotificationCenter.default.addObserver(
             forName: NSApplication.didChangeScreenParametersNotification,
             object: nil,
@@ -42,9 +45,13 @@ class ScreenObserver: ObservableObject {
                 self.screens = current
                 self.previousScreens = current
 
-                if UserDefaults.standard.bool(forKey: "autoSyncNewScreens"), let source = current.first {
-                    for screen in added {
-                        SharedWallpaperWindowManager.shared.syncWindow(to: screen, from: source)
+                if UserDefaults.standard.bool(forKey: "autoSyncNewScreens") {
+                    let srcID = AppState.shared.primaryScreenID
+                    let source = srcID.flatMap { NSScreen.screen(forDisplayID: $0) } ?? current.first
+                    if let src = source {
+                        for screen in added {
+                            SharedWallpaperWindowManager.shared.syncWindow(to: screen, from: src)
+                        }
                     }
                 }
             }
@@ -122,7 +129,7 @@ struct ContentView: View {
                 get: { !isMenuBarOnly },
                 set: { newValue in
                     isMenuBarOnly = !newValue
-                    AppDelegate.shared?.setDockIconVisible(newValue)
+                    AppAppearanceManager.shared.setDockIconVisible(newValue)
                 }
             ))
         }
@@ -162,6 +169,14 @@ struct ContentView: View {
         }
         .padding()
         .frame(maxHeight: .infinity)
+        .onAppear {
+            AppState.shared.primaryScreenID = selectedTabScreen?.dv_displayID
+            dlog("ContentView onAppear primary=\(String(describing: selectedTabScreen?.dv_localizedName))")
+        }
+        .onChange(of: selectedTabScreen) { newScreen in
+            dlog("selectedTabScreen changed \(newScreen?.dv_localizedName ?? "nil")")
+            AppState.shared.primaryScreenID = newScreen?.dv_displayID
+        }
     }
 }
 
@@ -328,6 +343,7 @@ struct SingleScreenView: View {
 
             if !NSScreen.screens.contains(screen) {
                 selectedTabScreen = NSScreen.screens.first
+                AppState.shared.primaryScreenID = selectedTabScreen?.dv_displayID
             }
         }
         .onAppear {
